t2685: split signature code into sibling module to fix qlty regression gate
The t2685 implementation added ~270 lines of signature-footer enforcement
code to quality-hooks.mjs, pushing the file into qlty's file-complexity
threshold. Result: PR #20307 passed the absolute-threshold gate (24 < 25)
but tripped the regression gate (base=23 head=24, +1 file-complexity on
quality-hooks.mjs).

Extract tryRepairSignature, checkSignatureFooterGate, and their helpers
into quality-hooks-signature.mjs. quality-hooks.mjs now re-exports them
via a 7-line ESM re-export. No behavioural change.

Verification:
- 32/32 signature footer gate tests pass
- 12/12 PATH shim tests pass
- 30/30 gh_wrapper auto-sig tests pass
- 106/106 full plugin suite pass
- qlty smells total: 23 (was 24 pre-split, below 25 threshold)
- qlty regression delta on quality-hooks.mjs: +0 (was +1)

Dead imports readFileSync and appendFileSync also removed from
quality-hooks.mjs since they were only used by the extracted signature code.
